first iteration of image CDN - minimum MVP
this works by storing image thumbnails the first time they're requested,
then serving the cached image in subsequent requests.

just rolling my own for the moment since off-the-shelf CDNs don't seem
quite suited to our requirements and i'd like to try the simplest
possible approach first.

i've plumbed this through to the frontend, though it's strictly opt-in
at the moment, and opting in causes a banner to appear on the home page
indicating that an experimental feature is being tested.

i'm expecting to roll this out to mpcautofill.github.io for beta testing
once the todo list below is all ticked off.

TODO:
* see if we can set up end-to-end tests with a full docker compose build
* set up web sockets to listen to google drive for changes to images
* set a maximum time-to-live for thumbnails and set up a cron job to
  clean them up every day (is daily the correct timing?)
  * this is a precaution in case any messages from google drive are
    dropped
* research what headers i need to set to ensure this plays nice with
  an actual CDN sitting in front of it